+++ server/src/engine/storage/v2/impls/mdl_journal.rs
@@ -64,6 +64,23 @@ use {
     },
 };
 
+#[cfg(test)]
+local! {
+    static BATCH_INFO: Vec<BatchInfo> = Vec::new();
+}
+
+#[cfg(test)]
+#[derive(Debug, PartialEq)]
+pub struct BatchInfo {
+    pub items_count: usize,
+    pub redundant_count: usize,
+}
+
+#[cfg(test)]
+pub fn get_last_batch_run_info() -> Vec<BatchInfo> {
+    local_mut!(BATCH_INFO, |info| core::mem::take(info))
+}
+
 pub type ModelDriver = BatchDriver<ModelDataAdapter>;
 impl ModelDriver {
     pub fn open_model_driver(
@@ -333,17 +350,10 @@ impl<'a> FullModel<'a> {
     pub fn new(model: &'a ModelData) -> Self {
         Self(model)
     }
-}
-
-impl<'a> JournalAdapterEvent<BatchAdapter<ModelDataAdapter>> for FullModel<'a> {
-    fn md(&self) -> u64 {
-        BatchType::Standard.dscr_u64()
-    }
-    fn write_direct(
+    fn write<const ZERO: bool>(
         self,
-        f: &mut TrackedWriter<<BatchAdapter<ModelDataAdapter> as RawJournalAdapter>::Spec>,
-        _: Rc<RefCell<BatchStats>>,
-    ) -> RuntimeResult<()> {
+        f: &mut TrackedWriter<ModelDataBatchAofV1>,
+    ) -> Result<(), crate::engine::fractal::error::Error> {
         let g = pin();
         let mut row_writer: RowWriter<'_> = RowWriter { f };
         let index = self.0.primary_index().__raw_index();
@@ -356,7 +366,14 @@ impl<'a> JournalAdapterEvent<BatchAdapter<ModelDataAdapter>> for FullModel<'a> {
         row_writer.write_row_global_metadata(self.0)?;
         for (key, row_data) in index.mt_iter_kv(&g) {
             let row_data = row_data.read();
-            row_writer.write_row_metadata(DataDeltaKind::Insert, row_data.get_txn_revised())?;
+            row_writer.write_row_metadata(
+                DataDeltaKind::Insert,
+                if ZERO {
+                    DeltaVersion::genesis()
+                } else {
+                    row_data.get_txn_revised()
+                },
+            )?;
             row_writer.write_row_pk(key)?;
             row_writer.write_row_data(self.0, &row_data)?;
         }
@@ -365,6 +382,19 @@ impl<'a> JournalAdapterEvent<BatchAdapter<ModelDataAdapter>> for FullModel<'a> {
             .f
             .dtrack_write(&current_row_count.u64_bytes_le())?;
         Ok(())
+    }
+}
+
+impl<'a> JournalAdapterEvent<BatchAdapter<ModelDataAdapter>> for FullModel<'a> {
+    fn md(&self) -> u64 {
+        BatchType::Standard.dscr_u64()
+    }
+    fn write_direct(
+        self,
+        f: &mut TrackedWriter<<BatchAdapter<ModelDataAdapter> as RawJournalAdapter>::Spec>,
+        _: Rc<RefCell<BatchStats>>,
+    ) -> RuntimeResult<()> {
+        self.write::<false>(f)
     }
 }
 
@@ -434,6 +464,20 @@ impl BatchStats {
     }
 }
 
+struct ModelConslidation<'a>(&'a ModelData);
+impl<'a> JournalAdapterEvent<BatchAdapter<ModelDataAdapter>> for ModelConslidation<'a> {
+    fn md(&self) -> u64 {
+        BatchType::Standard.dscr_u64()
+    }
+    fn write_direct(
+        self,
+        f: &mut TrackedWriter<<BatchAdapter<ModelDataAdapter> as RawJournalAdapter>::Spec>,
+        _: <BatchAdapter<ModelDataAdapter> as RawJournalAdapter>::CommitContext,
+    ) -> RuntimeResult<()> {
+        FullModel(self.0).write::<true>(f)
+    }
+}
+
 impl BatchAdapterSpec for ModelDataAdapter {
     type Spec = ModelDataBatchAofV1;
     type GlobalState = ModelData;
@@ -447,7 +491,14 @@ impl BatchAdapterSpec for ModelDataAdapter {
         writer: &mut ModelDriver,
         ctx: Self::FullSyncCtx<'a>,
     ) -> RuntimeResult<()> {
-        writer.commit_with_ctx(FullModel(ctx), BatchStats::new())
+        /*
+            a batch consolidation is our opportunity to fully reset version counters to genesis (+1 because of fetch add).
+            basically after the compaction "all events already happened" and the next event to happen will have ID 1
+        */
+        writer.commit_with_ctx(ModelConslidation(ctx), BatchStats::new())?;
+        ctx.delta_state()
+            .__set_delta_version(DeltaVersion::__new(1));
+        Ok(())
     }
     fn is_early_exit(event_type: &Self::EventType) -> bool {
         EventType::EarlyExit.eq(event_type)
@@ -539,6 +590,8 @@ impl BatchAdapterSpec for ModelDataAdapter {
         let m = gs;
         let mut real_last_txn_id = DeltaVersion::genesis();
         let mut redundant_records = 0;
+        #[cfg(test)]
+        let ev_count = batch_state.events.len();
         for DecodedBatchEvent { txn_id, pk, kind } in batch_state.events {
             match kind {
                 DecodedBatchEventKind::Insert(new_row)
@@ -665,6 +718,13 @@ impl BatchAdapterSpec for ModelDataAdapter {
         m.delta_state()
             .__set_delta_version(DeltaVersion::__new(real_last_txn_id.value_u64() + 1));
         heuristics.report_additional_redundant_records(redundant_records);
+        #[cfg(test)]
+        {
+            local_mut!(BATCH_INFO, |info| info.push(BatchInfo {
+                items_count: ev_count,
+                redundant_count: redundant_records
+            }))
+        }
         Ok(())
     }
 }